Dakota/kyb_review/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kyb"
version = "1.3.2"
description = "API para automatización de procesos Know-Your-Business en el sector bancario mexicano"
requires-python = ">=3.12, <3.14"
readme = "README.md"

dependencies = [
    # ── Azure AI ─────────────────────────────────────────
    "azure-ai-documentintelligence>=1.0.2",
    "azure-ai-textanalytics>=5.3.0",
    "azure-core>=1.35.0",

    # ── Web framework ────────────────────────────────────
    "fastapi[standard]>=0.116.1",
    "starlette>=0.47.2",
    "python-multipart>=0.0.20",
    "pydantic>=2.11.7",
    "email-validator>=2.2.0",

    # ── LLM / LangChain ─────────────────────────────────
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "openai>=0.27.0",

    # ── MCP ──────────────────────────────────────────────
    "fastmcp>=2.10.6",

    # ── PDF / OCR ────────────────────────────────────────
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pypdf>=5.9.0",
    "pytesseract>=0.3.13",
    "tesseract>=0.1.3",

    # ── NLP ──────────────────────────────────────────────
    "spacy>=3.8.7",
    "thefuzz[speedup]>=0.22.1",

    # ── Base de datos (async) ────────────────────────────
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",    "alembic>=1.13.0",
    # ── Utilidades ───────────────────────────────────────
    "python-dotenv>=1.0.0",
    "python-magic>=0.4.27",
    "rich>=14.0.0",
    "nest_asyncio>=1.6.0",


]

[project.optional-dependencies]
windows = [
    "python-magic-bin>=0.4.14",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["api*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
